+++ x/distribution/migrations/v4/migrate_test.go
@@ -8,14 +8,14 @@ import (
 	"cosmossdk.io/collections"
 	collcodec "cosmossdk.io/collections/codec"
 	storetypes "cosmossdk.io/store/types"
+	"cosmossdk.io/x/distribution"
+	v4 "cosmossdk.io/x/distribution/migrations/v4"
 
 	"github.com/cosmos/cosmos-sdk/crypto/keys/secp256k1"
 	"github.com/cosmos/cosmos-sdk/runtime"
 	"github.com/cosmos/cosmos-sdk/testutil"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	moduletestutil "github.com/cosmos/cosmos-sdk/types/module/testutil"
-	"github.com/cosmos/cosmos-sdk/x/distribution"
-	v4 "github.com/cosmos/cosmos-sdk/x/distribution/migrations/v4"
 )
 
 func TestMigration(t *testing.T) {